feat(add): Implement efficient add opcode
Implement the  _add  function in  monty.h  to efficiently handle the add opcode. The function takes a pointer to the stack and the line count as arguments. It first checks if the stack is empty or if the stack has less than two elements (top and next). If either condition is true, it prints an error message indicating that the stack is too short and sets the  status  variable to  EXIT_FAILURE . If the stack has at least two elements, it calculates the sum of the values of the top two elements by adding  ((*stack)->next->n)  and  ((*stack)->n) . It then calls the  pop  function to remove the top element from the stack. Finally, it updates the value of the new top element  ((*stack)->n)  with the calculated result.