feat(mul): Implement efficient mul opcode
Implement the  _mul  function in  monty.h  to efficiently handle the mul opcode. The function takes a pointer to the stack and the line count as arguments. It first checks if the stack is empty or if the stack has less than two elements (top and next). If either condition is true, it prints an error message indicating that the stack is too short and sets the  status  variable to  EXIT_FAILURE . If the stack has at least two elements, it calculates the product of the values of the second top element and the top element by multiplying  ((*stack)->next->n)  and  ((*stack)->n) . It then calls the  pop  function to remove the top element from the stack. Finally, it updates the value of the new top element  ((*stack)->n)  with the calculated result.