feat(rotr): Implement efficient rotr opcode
Implement the  _rotr  function in  monty.h  to efficiently handle the rotr opcode. The function takes a double pointer to the stack and the line count as arguments. It first checks if the stack is empty or contains only one element. If either condition is true, it returns. If the stack has more than one element, it creates two stack pointers,  bottom  and  prev , and initializes  bottom  to the stack pointer. It then traverses the stack to find the last node, which will become the new top element. It updates the  next  and  prev  pointers of the last node ( bottom ) and the previous node ( prev ) accordingly. It breaks the link between the previous top elements and the new top element by setting the  next  pointer of the previous top element to  NULL . Finally, it updates the stack pointer to point to the new top element.