chore(commit): Add efficient is_digit and isnumber functions
Add the  is_digit  and  isnumber  functions to the  monty.h  file. The  is_digit  function checks if a string is a digit. It returns 1 if the string is a digit and 0 if it is not. The  isnumber  function checks if a string is a number. It returns 1 if the string is a number and 0 if it is not.